// CMIS4-FE/src/app/modules/danh-muc-module/models/dmuc-nhan-cong/nhan-cong.model.ts
export interface nhancong {
    MA_DVIQLY: string;

    maVtu: string;

    tenVtu?: string;

    MA_LOAI_CPHI: string;

    DON_GIA?: number;

    DVI_TINH?: string;

    PHA?: number;
    STT_1PHA?: number;
    STT_3PHA?: number;
    TRU?: number;

    CONG_THUC?: string;

    CO_THUE?: number;
    TRANG_THAI?: number;
}